Remove MatchingProgressBar and update MatchingTitle in matching components
